Cache FDA alterations
