+++ core/src/main/java/org/mskcc/cbio/oncokb/cache/CacheFetcher.java
@@ -1,24 +1,21 @@
 package org.mskcc.cbio.oncokb.cache;
 
-import org.apache.commons.lang3.StringUtils;
 import org.mskcc.cbio.oncokb.apiModels.CuratedGene;
-import org.mskcc.cbio.oncokb.apiModels.annotation.AnnotateMutationByHGVSgQuery;
+import org.mskcc.cbio.oncokb.apiModels.FdaAlteration;
 import org.mskcc.cbio.oncokb.apiModels.annotation.AnnotationQueryType;
 import org.mskcc.cbio.oncokb.bo.OncokbTranscriptService;
 import org.mskcc.cbio.oncokb.genomenexus.GNVariantAnnotationType;
 import org.mskcc.cbio.oncokb.model.*;
 import org.mskcc.cbio.oncokb.util.*;
 import org.oncokb.oncokb_transcript.ApiException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
+import java.util.stream.Collectors;
 
 import static org.mskcc.cbio.oncokb.Constants.DEFAULT_REFERENCE_GENOME;
+import static org.mskcc.cbio.oncokb.model.InferredMutation.ONCOGENIC_MUTATIONS;
 
 @Component
 public class CacheFetcher {
@@ -174,6 +171,100 @@ public class CacheFetcher {
     @Cacheable(cacheResolver = "generalCacheResolver")
     public Gene findGeneBySymbol(String symbol) throws ApiException {
         return this.oncokbTranscriptService.findGeneBySymbol(symbol);
+    }
+
+    @Cacheable(cacheResolver = "generalCacheResolver", key = "'all'")
+    public Set<FdaAlteration> getAllFdaAlterations() {
+        Set<Gene> genes = CacheUtils.getAllGenes();
+        Set<FdaAlteration> alterations = new HashSet<>();
+        Map<Alteration, Map<String, Set<FdaAlteration>>> resultMap = new HashMap<>();
+        for (Gene gene : genes) {
+            for (ClinicalVariant clinicalVariant : MainUtils.getClinicalVariants(gene)) {
+                LevelOfEvidence level = LevelOfEvidence.getByLevel(clinicalVariant.getLevel());
+                if (level == null || !LevelUtils.getTherapeuticLevels().contains(level)) {
+                    continue;
+                }
+                Set<Alteration> mappedAlterations = new HashSet<>();
+                if (clinicalVariant.getVariant().getAlteration().equals(ONCOGENIC_MUTATIONS.getVariant())) {
+                    for (BiologicalVariant annotatedAlt : MainUtils.getBiologicalVariants(gene)) {
+                        Oncogenicity oncogenicity = Oncogenicity.getByEffect(annotatedAlt.getOncogenic());
+                        if (MainUtils.isOncogenic(oncogenicity)) {
+                            mappedAlterations.add(annotatedAlt.getVariant());
+                        }
+                    }
+                } else {
+                    mappedAlterations.add(clinicalVariant.getVariant());
+                }
+                for (Alteration alteration : mappedAlterations) {
+                    if (!resultMap.containsKey(alteration)) {
+                        resultMap.put(alteration, new HashMap<>());
+                    }
+                    for (TumorType tumorType : clinicalVariant.getCancerTypes()) {
+                        FdaAlteration fdaAlteration = new FdaAlteration();
+                        String cancerTypeName = TumorTypeUtils.getTumorTypeName(tumorType);
+                        fdaAlteration.setAlteration(alteration);
+                        fdaAlteration.setLevel(convertToFdaLevel(LevelOfEvidence.getByLevel(clinicalVariant.getLevel()), clinicalVariant));
+                        fdaAlteration.setCancerType(cancerTypeName);
+                        if (!resultMap.get(alteration).containsKey(cancerTypeName)) {
+                            resultMap.get(alteration).put(cancerTypeName, new HashSet<>());
+                        }
+                        resultMap.get(alteration).get(cancerTypeName).add(fdaAlteration);
+                    }
+                }
+            }
+        }
+
+        for (Map.Entry<Alteration, Map<String, Set<FdaAlteration>>> alterationMap : resultMap.entrySet()) {
+            for (Map.Entry<String, Set<FdaAlteration>> cancerTypeMap : alterationMap.getValue().entrySet()) {
+                FdaAlteration pickedFdaAlt = cancerTypeMap.getValue().iterator().next();
+                if (cancerTypeMap.getValue().size() > 1) {
+                    Optional<FdaAlteration> level2 = cancerTypeMap.getValue().stream().filter(fdaAlteration -> fdaAlteration.getLevel().equals(FDA_L_2)).findAny();
+                    if (level2.isPresent()) {
+                        pickedFdaAlt = level2.get();
+                    }
+                }
+                alterations.add(pickedFdaAlt);
+            }
+        }
+        return alterations;
+    }
+
+    final String FDA_L_2 = "FDAx2";
+    final String FDA_L_3 = "FDAx3";
+    private String convertToFdaLevel(LevelOfEvidence level, ClinicalVariant clinicalVariant) {
+        if (level == null) {
+            return "";
+        }
+        switch (level) {
+            case LEVEL_1:
+            case LEVEL_R1:
+                return FDA_L_2;
+            case LEVEL_2:
+                if (specialFdaL3(clinicalVariant)) {
+                    return FDA_L_3;
+                } else {
+                    return FDA_L_2;
+                }
+            case LEVEL_3A:
+            case LEVEL_4:
+            case LEVEL_R2:
+                return FDA_L_3;
+            default:
+                return "";
+        }
+    }
+
+    private boolean specialFdaL3(ClinicalVariant clinicalVariant) {
+        List<String> specialList = Arrays.asList(new String[]{"ERBB2&Oncogenic Mutations&Non-Small Cell Lung Cancer&Ado-Trastuzumab Emtansine", "EGFR&A763_Y764insFQEA&Non-Small Cell Lung Cancer&Erlotinib", "ALK&Fusions&Inflammatory Myofibroblastic Tumor&Crizotinib", "ALK&Fusions&Inflammatory Myofibroblastic Tumor&Ceritinib", "BRAF&V600E&Ganglioglioma, Pleomorphic Xanthoastrocytoma, Pilocytic Astrocytoma&Cobimetinib+Vemurafenib,", "BRAF&V600E&Ganglioglioma, Pleomorphic Xanthoastrocytoma, Pilocytic Astrocytoma&Trametinib+Dabrafenib"});
+
+        String separator = "&";
+        List<String> queryParts = new ArrayList<>();
+        queryParts.add(clinicalVariant.getVariant().getGene().getHugoSymbol());
+        queryParts.add(clinicalVariant.getVariant().getAlteration());
+        queryParts.add(TumorTypeUtils.getTumorTypesName(clinicalVariant.getCancerTypes()));
+        queryParts.add(String.join("+", clinicalVariant.getDrug().stream().sorted().collect(Collectors.toList())));
+        String query = String.join(separator, queryParts);
+        return specialList.contains(query);
     }
 
     @Cacheable(
